ENH: Update UnitTestCommon to have a .cpp file for implementations
This is handy because on macOS when debugging the requirements for the decompression never passes even though it actually does work. This allows the developer to quickly comment out that line in order to debug.
